+++ packages/frameworks/vue/src/popover/tests/popover.test.tsx
@@ -57,7 +57,7 @@ describe('Popover', () => {
     expect(screen.queryByText('title')).not.toBeVisible()
   })
 
-  it.skip('should be able to lazy mount', async () => {
+  it('should be able to lazy mount', async () => {
     render(ComponentUnderTest, {
       props: {
         lazyMount: true,
@@ -73,7 +73,7 @@ describe('Popover', () => {
     expect(screen.getByTestId('positioner')).toBeInTheDocument()
   })
 
-  it.skip('should not have aria-controls if lazy mounted', async () => {
+  it('should not have aria-controls if lazy mounted', async () => {
     render(ComponentUnderTest, {
       props: {
         lazyMount: true,
@@ -81,5 +81,22 @@ describe('Popover', () => {
     })
 
     expect(screen.getByRole('button', { name: 'click me' })).not.toHaveAttribute('aria-controls')
+  })
+
+  it('should lazy mount and unmount on exit', async () => {
+    render(ComponentUnderTest, {
+      props: {
+        lazyMount: true,
+        unmountOnExit: true,
+      },
+    })
+
+    expect(screen.queryByTestId('positioner')).not.toBeInTheDocument()
+
+    await user.click(screen.getByRole('button', { name: 'click me' }))
+    expect(screen.getByTestId('positioner')).toBeInTheDocument()
+
+    await user.click(screen.getByRole('button', { name: 'close' }))
+    expect(screen.queryByTestId('positioner')).not.toBeInTheDocument()
   })
 })